Title: ab곱 n번 구하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240901/ab곱 n번 구하기/a-multiple-b-n-times.c
@@ -9,7 +9,7 @@ int main() {
         int a, b;
         scanf("%d %d", &a, &b); 
 
-        long long p = 1;
+        int p = 1;
         for (int j = a; j <= b; j++) {
             p *= j;  
         }
